fixed -allow-debug for snp guests
For some reasons the launch script did not honor the
`-allow-debug` flag with sev snp guests. Strangely
this still worked using the 5.14 kernel for SEV-Step
(i.e. we where able to use debug commands).
With the 5.19 kernel this stopped working (which is the correct
behavior, since we did not set the policy bit)